refactor: Reorganize files and components
