+++ frontend/src/components/common/layout.tsx
