+++ frontend/src/components/common/navbar.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 import Image from "next/image";
 
 export default function Navbar() {
