+++ frontend/src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import '@/styles/globals.css'
 import type { AppProps } from 'next/app'
-import Layout from '../components/layout'
+import Layout from '../components/common/layout'
 import { Noto_Sans } from 'next/font/google'
 
 const notoSans = Noto_Sans({
